requirements.txt
alabaster==0.7.16
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asteval==1.0.1
astroid==3.2.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
bandit==1.7.9
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
bokeh==3.4.2
-e git+https://github.com/BonnerLab/bonner-libraries.git@66fd28e24172efdf05084d662a7c0a2e2e026528#egg=bonner_libraries
boto3==1.34.144
botocore==1.34.144
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
cftime==1.6.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dask==2024.7.0
datashader==0.16.3
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
docutils==0.21.2
emcee==3.1.6
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
graphviz==0.20.3
h11==0.14.0
holoviews==1.19.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanize==4.10.0
idna==3.7
imagesize==1.4.1
importlib_metadata==8.0.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
jupytext==1.16.3
kiwisolver==1.4.5
lightning-utilities==0.11.4
line_profiler==4.1.3
linkify-it-py==2.0.3
llvmlite==0.43.0
lmfit==1.3.1
locket==1.0.0
loguru==0.7.2
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
multipledispatch==1.0.0
mypy==1.10.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.1.post1
networkx==3.3
nibabel==5.2.1
nilearn==0.10.4
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.0
osfclient==0.0.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
panel==1.4.4
param==2.1.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==0.5.6
pbr==6.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pyct==0.5.0
Pygments==2.18.0
pylint==3.2.5
pynndescent==0.5.13
pyparsing==3.1.2
pyproject-api==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyviz_comms==3.0.2
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
snakeviz==2.2.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
statsmodels==0.14.2
stevedore==5.2.0
svgutils==0.3.4
sympy==1.13.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.7
tinycss2==1.3.0
tomli==2.0.1
tomlkit==0.13.0
toolz==0.12.1
torch==2.3.1
torchdata==0.7.1
torchlens==0.1.17
torchmetrics==1.4.0.post0
torchvision==0.18.1
tornado==6.4.1
tox==4.16.0
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
uncertainties==3.2.2
-e git+https://github.com/zche377/universal_dimensions.git@a33efdeb3a65b6e3ee5437cc4cb3d9a7a3914dbf#egg=universal_dimensions
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wget==3.2
widgetsnbextension==4.0.11
xarray==2024.6.0
xyzservices==2024.6.0
zenodo-get==1.6.1
zipp==3.19.2
